modularize madara config
