+++ crates/madara/src/config/global_config.rs
@@ -22,43 +22,11 @@ use serde::{Deserialize, Serialize};
 use crate::constants::DEFAULT_LOCAL_CONFIG_FILE;
 
 use super::{
-    constants::{
-        DEFAULT_ATLANTIC_URL, MADARA_APP_CHAIN_ID, MADARA_BLOCK_TIME, MADARA_CHAIN_NAME,
-        MADARA_LATEST_PROTOCOL_VERSION, MADARA_NATIVE_FEE_TOKEN_ADDRESS,
-        MADARA_PARENT_FEE_TOKEN_ADDRESS, MADARA_PENDING_BLOCK_UPDATE_TIME,
-    },
+    constants::DEFAULT_ATLANTIC_URL,
     eth_wallet::EthWallet,
     l1_config::{self, L1Configuration},
+    madara::MadaraConfiguration,
 };
-
-#[derive(Debug, Serialize, Deserialize, Clone)]
-pub struct MadaraConfiguration {
-    pub chain_name: String,
-    pub app_chain_id: String,
-    pub native_fee_token_address: String,
-    pub parent_fee_token_address: String,
-    pub latest_protocol_version: String,
-    pub block_time: String,
-    pub pending_block_update_time: String,
-    pub gas_price: u64,
-    pub blob_gas_price: u64,
-}
-
-impl Default for MadaraConfiguration {
-    fn default() -> Self {
-        Self {
-            chain_name: MADARA_CHAIN_NAME.to_string(),
-            app_chain_id: MADARA_APP_CHAIN_ID.to_string(),
-            native_fee_token_address: MADARA_NATIVE_FEE_TOKEN_ADDRESS.to_string(),
-            parent_fee_token_address: MADARA_PARENT_FEE_TOKEN_ADDRESS.to_string(),
-            latest_protocol_version: MADARA_LATEST_PROTOCOL_VERSION.to_string(),
-            block_time: MADARA_BLOCK_TIME.to_string(),
-            pending_block_update_time: MADARA_PENDING_BLOCK_UPDATE_TIME.to_string(),
-            gas_price: 0,
-            blob_gas_price: 0,
-        }
-    }
-}
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
 pub struct OrchestratorConfiguration {
@@ -116,36 +84,7 @@ impl Config {
         EthWallet::init(&mut local_template)?;
 
         // Madara configuration
-        logger::new_empty_line();
-        logger::note(
-            "Madara configuration",
-            "You'll need to setup all the parameters related to Madara",
-        );
-        let chain_name = Prompt::new("Enter the name of the Madara chain (e.g., MyChain)")
-            .default(&local_template.madara.chain_name)
-            .ask();
-        let app_chain_id = Prompt::new("Enter the Madara chain ID (e.g., MADARA_DEVNET)")
-            .default(&&local_template.madara.app_chain_id)
-            .ask();
-        let block_time = Prompt::new("Enter the block time for Madara (in seconds, e.g., 15s)")
-            .default(&local_template.madara.block_time)
-            .validate_interactively(validate_time_with_unit)
-            .ask();
-        let gas_price = Prompt::new("Enter the gas price for Madara (in Gwei, e.g., 20)")
-            .default(&local_template.madara.gas_price.to_string())
-            .validate_interactively(validate_u64)
-            .ask::<u64>();
-
-        let blob_gas_price = Prompt::new("Enter the blob gas price for Madara (in Gwei, e.g., 5)")
-            .default(&local_template.madara.blob_gas_price.to_string())
-            .validate_interactively(validate_u64)
-            .ask::<u64>();
-
-        local_template.madara.chain_name = chain_name;
-        local_template.madara.app_chain_id = app_chain_id;
-        local_template.madara.block_time = block_time;
-        local_template.madara.gas_price = gas_price;
-        local_template.madara.blob_gas_price = blob_gas_price;
+        MadaraConfiguration::init(&mut local_template)?;
 
         // Orchestrator configuration
         logger::new_empty_line();
